Rename ClientAutoMapperProfile to IdentityServerAutoMapperProfile and move to the Volo.Abp.IdentityServer namespace
